RAG/rag-with-vectorsearch-2.0/rag_with_vectorsearch_2_0/requirements.txt
# Agent Framework
google-adk==1.22.0
python-dotenv==1.2.1
google-auth>=2.47.0

# Vertex AI SDK (for Vector Search 2.0 and Embeddings)
google-cloud-aiplatform>=1.33.0
google-cloud-vectorsearch==0.2.0
